feat: refactor tent detail page and add control/event history sections
